moved from float to double

+++ src/Vec3/vec3.cpp
@@ -17,11 +17,11 @@ vec3 vec3::operator-() const {
 	return vec3(-1*axis[0], -1*axis[1], -1*axis[2]);
 }
 
-float vec3::operator[](int i) const{
+double vec3::operator[](int i) const{
 	return axis[i];
 }
 
-float& vec3::operator[](int i){
+double& vec3::operator[](int i){
 	return axis[i];
 }
 
@@ -32,25 +32,25 @@ vec3& vec3::operator+=(const vec3 &v){
 	return *this;	
 }
 
-vec3& vec3::operator*=(const float t){
+vec3& vec3::operator*=(const double t){
 	axis[0]*=t;
 	axis[1]*=t;
 	axis[2]*=t;
 	return *this;
 }
 
-vec3& vec3::operator/=(const float t){
+vec3& vec3::operator/=(const double t){
 	axis[0]/=t;
 	axis[1]/=t;
 	axis[2]/=t;
 	return *this;
 }
 
-float vec3::length() const{
+double vec3::length() const{
 	return std::sqrt(length_squared());
 }
 
-float vec3::length_squared() const{
+double vec3::length_squared() const{
 	return axis[0]*axis[0] +  axis[1]*axis[1] + axis[2]*axis[2] ;
 }
 
@@ -72,15 +72,15 @@ inline vec3 operator*(const vec3 &u, const vec3 &v){
 	return vec3(u.axis[0] * v.axis[0], u.axis[1] * v.axis[1],u.axis[2] * v.axis[2]);
 }
 
-inline vec3 operator*(const float t,const vec3 &v){
+inline vec3 operator*(const double t,const vec3 &v){
 	return vec3(v.axis[0] * t, v.axis[1] * t,v.axis[2] * t);
 }
 
-inline vec3 operator*(const vec3 &u,const float t){
+inline vec3 operator*(const vec3 &u,const double t){
 	return u * t;
 }
 
-inline vec3 operator/(const vec3 &u,const float t){
+inline vec3 operator/(const vec3 &u,const double t){
 	return (1/t) * u;
 }
